Name the two things the tables kept asserting about
The universe is defined on every page a table renders on, and a metadata row
either describes representatives or doesn't. Both were spelled out inline at
each use site, the second one inside a useMemo in the row component.
