+++ react/src/components/congressional-table/model.ts
@@ -1,3 +1,6 @@
+import { assert } from '../../utils/defensive'
+import { MetadataStatValue } from '../load-article'
+
 export interface CongressionalRepresentativeEntry {
     representative: {
         name: string
@@ -12,6 +15,35 @@ export interface CongressionalRepresentativeEntry {
 export interface CongressionalColumnData {
     longname: string
     representatives: CongressionalRepresentativeEntry[]
+}
+
+/**
+ * Returns undefined for rows that aren't a representatives metadata row. Takes the
+ * row structurally so it works with both ArticleRow and StatisticCellRenderingInfo.
+ */
+export function congressionalDataForRow(
+    row: { kind: 'statistic' } | { kind: 'metadata', statval: MetadataStatValue },
+    longname: string,
+): CongressionalColumnData | undefined {
+    if (row.kind !== 'metadata' || typeof row.statval === 'string') {
+        return undefined
+    }
+    return {
+        longname,
+        representatives: row.statval.representatives.map((r): CongressionalRepresentativeEntry => {
+            assert(r.representative.name !== undefined && r.representative.name !== null, 'representative name missing')
+            return {
+                representative: {
+                    name: r.representative.name,
+                    wikipediaPage: r.representative.wikipediaPage ?? undefined,
+                    party: r.representative.party ?? undefined,
+                },
+                districtLongname: r.districtLongname,
+                startTerm: r.startTerm,
+                endTerm: r.endTerm,
+            }
+        }),
+    }
 }
 
 export interface CongressionalDisplayHeaderSpaceRow {
