+++ react/src/stat/StatisticPanelTable.tsx
@@ -7,10 +7,9 @@ import { CellSpec, SuperHeaderSpec, TableContents } from '../components/supertab
 import { ColumnIdentifier } from '../components/table'
 import { Navigator } from '../navigation/Navigator'
 import { useColors } from '../page_template/colors'
-import { useUniverse } from '../universe'
+import { useDefinedUniverse } from '../universe'
 import { orderNonNan } from '../urban-stats-script/constants/table'
 import { TypeEnvironment } from '../urban-stats-script/types-values'
-import { assert } from '../utils/defensive'
 import { reifyString } from '../utils/human-readable-name'
 import { sanitize } from '../utils/paths'
 
@@ -71,8 +70,7 @@ export function StatisticPanelTable({ view, stat, data, set, tableRef, loading, 
 
     const columnWidth = (100 - widthLeftHeader) / (data.table.length === 0 ? 1 : data.table.length)
 
-    const currentUniverse = useUniverse()
-    assert(currentUniverse !== undefined, 'no universe')
+    const currentUniverse = useDefinedUniverse()
 
     const onlyColumns: ColumnIdentifier[] = data.hideOrdinalsPercentiles ? ['statval', 'statval_unit'] : ['statval', 'statval_unit', 'statistic_ordinal', 'statistic_percentile']
